Adding two other files with a LCP analysis
I also wanted to test out a more complicated function; LCP analysis is something that could reasonable be utilized in a bumble bee foraging model, as linear features on the landscape tend to have an impact on foraging patterns, so deriving an LCP and evaluating the linearity of the path (not done here) could be useful. I'm not sure this will work, as I kind of had to ad hoc the json, but wanted to see!